Check pipeliner exit status to determine whether or not to delete the RUNNING_FILE

+++ src/relion/cryolo_relion_it/cryolo_relion_it.py
@@ -3604,8 +3604,9 @@ def run_pipeline(opts):
                     # if there isn't a new batch then check if the preprocessing is still running
                     preprocstatus = preproc.poll()
                     if preprocstatus is not None:
-                        os.remove(RUNNING_FILE)
-                        return False
+                        if preprocstatus != 0:
+                            os.remove(RUNNING_FILE)
+                        return
                     if CheckForExit():
                         return
                     # The following prevents checking the particles.star file too often
